Minor updates from Sean
Separated files into src and bin directories. I just figured out how to
compile Java into a directory separate from the source directory. You
use the -d parameter followed by the directory to compile to, which I
thought was useful. If you don’t think so, we can switch it back. Also,
just added constants in the Coordinate class to represent the possible
types. It made it easier for me to read and understand the code.